Remove unlawful instance
Line 48 of Data.Algebra.Boolean.Negable says that instances "must obey the double negation law: `not (not x) = x`". The provided instance for `Maybe` does not: if `x != mempty` then

    not (not (Just x)) = not Nothing = Just mempty

It is possible to give some lawful instances involving Maybe, like

    instance Negable (Maybe ())

but it is not clear that any of them are suitably canonical or suitably interesting.